Deleted extra file and put new code into existing file
